refactor(config): split config loading into helpers
