+++ .changeset/refactor-config-helpers.md
@@ -1,0 +1,5 @@
+---
+'@lapidist/design-lint': patch
+---
+
+refactor config loading into separate helpers
